% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{format_error}
\alias{format_error}
\title{Format error}
\usage{
format_error(simulation_error = NULL, alph = alph)
}
\arguments{
\item{simulation_error}{Simulation errors obtained from storing the results of sim_error() in MLAM.}

\item{alph}{Vector of alpha terms (L1 regularization parameter).}
}
\value{
Simulation error data.frame.
}
\description{
This function generates reformats the simulation error data from a list to a data.frame.
}
\examples{
simulate_error = format_error(simulation_error,alph)
}
