% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{get_annotations}
\alias{get_annotations}
\title{Format simulation annotations}
\usage{
get_annotations(sim = NULL, J = NULL)
}
\arguments{
\item{sim}{simulations.}

\item{J}{Jacobian matrix of parameters}
}
\value{
list(sim,J)
}
\description{
This function formats the final simulation data. This function is utilized by sim_best().
}
\examples{
best_params = select_best(simulate_error)
}
