% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{get_init_condits}
\alias{get_init_condits}
\title{Initial conditions}
\usage{
get_init_condits(datHMF = datHMF)
}
\arguments{
\item{datHMF}{The data matrix with scaled measurements measurements. Scaled center measurements are provided for 
each measurement/annotation (e.g., gene/organ) combination. This format is produced by scale_zeroOne(). Defaults to datHMF.}
}
\value{
An initial condition vector.
}
\description{
This function extracts the initial conditions for the dynamic variables corresponding to measurements.
}
\examples{
ic = get_init_condits(datHMF)
}
