% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m-value_functions.R
\name{get_M_indices}
\alias{get_M_indices}
\title{m-value information}
\usage{
get_M_indices(N_mRanges, mvals = NULL, Msc = 2, mr = NULL)
}
\arguments{
\item{N_mRanges}{The number of m-value ranges to use.}

\item{mvals}{A set of m-values, the object generated by get_M_vals(). Defaults to NULL.}

\item{Msc}{A scale factor that determines the M value (M = max(m)) corresponding to the largest range of m-values. 
Mmax <- ceiling(Msc * Np / 2) where Np = number of parameters.
Defaults to 2.}

\item{mr}{Indices for a specific m-value range.}
}
\value{
A List with m-value information.
}
\description{
Get indices in the vector of m-values and select which m-value ranges to use
}
\examples{
mRange_best = get_M_indices(N_mRanges,mvals,Msc,indM_best)$mRange
}
