% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m-value_functions.R
\name{get_M_vals}
\alias{get_M_vals}
\title{Select M ranges}
\usage{
get_M_vals(datHMF = datHMF, Mall = NULL, Msc = 2)
}
\arguments{
\item{datHMF}{The data matrix with scaled measurements measurements. Scaled center measurements are provided for 
each measurement/annotation (e.g., gene/organ) combination. This format is produced by scale_zeroOne(). Defaults to datHMF.}

\item{Mall}{A set of m-values. Defaults to NULL.}

\item{Msc}{A scale factor that determines the M value (M = max(m)) corresponding to the largest range of m-values. 
Mmax <- ceiling(Msc * Np / 2) where Np = number of parameters.
Defaults to 2.}
}
\value{
A List with the minimum number of modulating functions (Mmin), the maximum number of modulating functions (Mmax),
and the entire set of modulating function m-values (Mall).
}
\description{
Select a set of values for M = max(m)
}
\examples{
mvals = get_M_vals(datHMF, Msc=2)
}
