% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{get_rateConst}
\alias{get_rateConst}
\title{Get rate constants}
\usage{
get_rateConst(fit = NULL, nlambda = NULL)
}
\arguments{
\item{fit}{The beta values from the fit object generated by glmnet().}

\item{nlambda}{The number of lambda values.}
}
\value{
The rate constant matrix.
}
\description{
This function formats the rate constant data after the fir has been implemented using glmnet() from the glmnet library. 
Sometimes glmnet() produces less fits than requested based on the nlambda parameter. In such cases, get_rateConst() 
pads the parameter matrix with zeros such that the size of the matrix reflects the selection of nlambda.
}
\examples{
rateConst = get_rateConst(fit$beta,nlambda)
}
\references{
\url{https://www.jstatsoft.org/article/view/v033i01}
}
