% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hartley_functions.R
\name{getXY}
\alias{getXY}
\title{X and Y data}
\usage{
getXY(datHMF = datHMF, Mall = NULL, Msc = 2, n = 1, Harg0 = NULL)
}
\arguments{
\item{datHMF}{The data matrix with scaled measurements measurements. Scaled center measurements are provided for 
each measurement/annotation (e.g., gene/organ) combination. This format is produced by scale_zeroOne(). Defaults to datHMF.}

\item{Mall}{A set of m-values. Defaults to NULL.}

\item{Msc}{A scale factor that determines the M value (M = max(m)) corresponding to the largest range of m-values. 
Mmax <- ceiling(Msc * Np / 2) where Np = number of parameters.}

\item{n}{Highest order of the system (n=1 for the first degree system that this analysis was designed for).}

\item{Harg0}{If omega_0 is equal to zero (x=0), set omega_0 to Harg0 (w=Harg0).}
}
\value{
The X and Y transformations obtained using Equations 4-6 for all m-value ranges
}
\description{
This function function for getting the X and Y data (Equation 10, Anderson et al., 2017).
}
\examples{
XY = getXY(datHMF,Mall=Mall,Msc=Msc,n=n,Harg0=Harg0)
}
\references{
\url{http://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1005627}
}
