% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hartley_functions.R
\name{Hcompute}
\alias{Hcompute}
\title{The continuous Hartley transform}
\usage{
Hcompute(w = NULL, t = NULL, x = NULL, Harg0 = NULL)
}
\arguments{
\item{w}{This is omega_0 (Equation 4, Anderson et al., 2017). Defaults to NULL.}

\item{t}{Vector of measurement times. Defaults to NULL.}

\item{x}{Vector of measurement values. Defaults to NULL.}

\item{Harg0}{If omega_0 is equal to zero (x=0), set omega_0 to Harg0 (w=Harg0). Defaults to NULL.}
}
\value{
A vector with the transformation of the measurement data into the frequency domain.
}
\description{
This function performs linear interpolation between data points and implements the continuous Hartley transform
}
\examples{
This function is called by HMF1compute() and HMFcompute(). This function is not recommended to be used in isolation.
Hcompute(omega, time, expr, H0)
}
\references{
\url{http://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1005627}
}
