% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hartley_functions.R
\name{HMF1compute}
\alias{HMF1compute}
\title{HMF spectral component of the 1st derivative}
\usage{
HMF1compute(meds, m, W0, n, Harg0)
}
\arguments{
\item{meds}{Matrix with the first column containing measurement times and the second column containing the corresponding measurements. Defaults to NULL.}

\item{m}{The m value.}

\item{W0}{The omega_0 term.}

\item{n}{Highest order of the system (n=1 for the first degree system that this analysis was designed for).}

\item{Harg0}{If omega_0 is equal to zero (x=0), set omega_0 to Harg0 (w=Harg0).}
}
\value{
The spectral component data.
}
\description{
This function computes HMF spectral component of the 1st derivative of x (dx/dt = x', Equation 6, Anderson et al., 2017).
}
\examples{
This function is not recommended to be used in isolation.
HMF1compute(data, m, omega, n, H0)
}
\references{
\url{http://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1005627}
}
