% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMFmethod_mainFunction.R
\name{HMF_fit}
\alias{HMF_fit}
\title{System identification}
\usage{
HMF_fit(datHMF = NULL, N_mRanges = 10, alpha = NULL, nlambda = NULL,
  Mall = NULL, Msc = 2, n = 1, dt = NULL, epsilon = 10^(-10),
  Harg0 = 1e-06)
}
\arguments{
\item{datHMF}{The data matrix with scaled measurements measurements. Scaled center measurements are provided for 
each measurement/annotation (e.g., gene/organ) combination. This format is produced by scale_zeroOne(). Defaults to datHMF.}

\item{N_mRanges}{The number of m-value sets to evaluate.}

\item{alpha}{The L1 regularization terms for constraining the regression. Defaults to seq(0,1,0.2).}

\item{nlambda}{The number of L2 regularization terms for consideration. Defaults to 10.}

\item{Mall}{A set of m-values. Defaults to NULL.}

\item{Msc}{A scale factor that determines the M value (M = max(m)) corresponding to the largest range of m-values. 
Mmax <- ceiling(Msc * Np / 2) where Np = number of parameters.}

\item{n}{Highest order of the system (n=1 for the first degree system that this analysis was designed for).}

\item{dt}{Simulation time step}

\item{epsilon}{Term for computing simulation error. Defaults to 10^(-10).}

\item{Harg0}{If omega_0 is equal to zero (x=0), set omega_0 to Harg0 (w=Harg0).}
}
\value{
Error data for all simulations, simulation data for the best simulation, a parameter matrix for the best simulation.
}
\description{
This function performs system identification to infer the connectivity and dynamics of a 
linear network model (Anderson et al., 2017).
}
\examples{
err0 = HMF_fit(datHMF,N_mRanges=2,dt=0.1,alpha=c(0.2,0.6))
}
\references{
\url{http://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1005627}
}
