% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiorgan.R
\docType{data}
\name{multiorgan}
\alias{multiorgan}
\title{Multi-organ gene expression data}
\format{An object of class \code{data.frame} with 82 rows and 24 columns.}
\usage{
data(multiorgan)
}
\description{
Data are from Anderson et al., 2017, Plos Computational Biology.
}
\examples{
data(multiorgan)
times <- unique(multiorgan[,1])
organs <- unique(multiorgan[,2])
genes = colnames(multiorgan)[2:ncol(multiorgan)]
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{http://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1005627})
}
\keyword{datasets}
