% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_functions.R
\name{scale_zeroOne}
\alias{scale_zeroOne}
\title{Normalization 0,1}
\usage{
scale_zeroOne(mat, center = "mean")
}
\arguments{
\item{mat}{Matrix with the first column containing measurement times and the second column 
containing annotations (organs). The third and subsequent columns contain measurement data.}

\item{center}{The metric used to compute the center, options include "mean" and "median"}
}
\value{
The data matrix with scaled measurements measurements. Scaled center measurements are provided for 
each measurement/annotation (e.g., gene/organ) combination. This is the data format required for HMF().
}
\description{
This function normalizes input data to the range (0,1).
}
\examples{
datHMF = scale_zeroOne(matrix, center="mean")
}
