% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{select_best}
\alias{select_best}
\title{Best simulation}
\usage{
select_best(simulate_error = NULL)
}
\arguments{
\item{simulate_error}{simulation error data frame generated by format_error().}
}
\value{
Best simulation parameters.
}
\description{
This function gets the parameters (lambda, alpha, m-values) for the best simulation based on 
the minumum of the simulation error.
}
\examples{
best_params = select_best(simulate_error)
}
