% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{sim_error}
\alias{sim_error}
\title{Simulate the dynamic model and record the agreement with the data.}
\usage{
sim_error(datHMF = NULL, Nr = NULL, tlen = NULL, rateConst = NULL,
  ic = NULL, dt = NULL, epsilon = NULL, numPar = NULL)
}
\arguments{
\item{datHMF}{The data matrix with scaled measurements measurements. Scaled center measurements are provided for 
each measurement/annotation (e.g., gene/organ) combination. This format is produced by scale_zeroOne(). Defaults to datHMF.}

\item{Nr}{Number of annotations (e.g., organs).}

\item{tlen}{Duration of simulation time.}

\item{rateConst}{Rate constant matrix.}

\item{ic}{Initial conditions.}

\item{dt}{Time step.}

\item{epsilon}{Term for computing error.}

\item{numPar}{Number of parameters.}
}
\value{
Simulation errors.
}
\description{
This function generates simulations and corresponding errors.
}
\examples{
MLAM[indLam,5] <- sim_error(datHMF,Nr,tlen,rateConst,ic,dt,epsilon,numPar)
}
\references{
\url{https://www.jstatsoft.org/article/view/v033i09}
}
