% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{sim_plots}
\alias{sim_plots}
\title{Plot simulation data}
\usage{
sim_plots(negddCt = NULL, Sims = NULL)
}
\arguments{
\item{negddCt}{Raw dara file.}

\item{Sims}{Simulation result.}
}
\description{
This function plots simulation data
}
\examples{
sim_plots(multiorgan, result$best_sim)
}
